metadata: Enable built time override in metadata
With the built time being part of an EIF's metadata we are not able to
reproducibly build a byte-for-byte identical image file from the same
sources.

Allow to optionally provide a build time override as DateTime<Utc>.
Without providing a build time override the current time is taken as has
always been the case.

Signed-off-by: Leonard Foerster <foersleo@amazon.de>